Refactor next previous hook
